airflow/requirements.txt
apache-airflow-providers-google
google-cloud-bigquery
pandas
dbt-bigquery